Исправлены ошибки и баги
