+++ .env
@@ -1,1 +1,1 @@
-API_PATH=http://127.0.0.1:9000/api
+API_PATH=http://127.0.0.1:8000/api
