--- src/UI/judgement/sass/groupStage.scss
@@ -1,3 +1,0 @@
-.showTeams{
-    align-items: center;
-}
