+++ src/UI/judgment/group-stage/sass/groupStage.scss
@@ -1,0 +1,18 @@
+.showTeams{
+    align-items: center;
+}
+.getMatches{
+    width: '100%'
+}
+.matchRow{
+    max-width: '100vw';
+    overflow-x: 'auto';
+    overflow-y: 'auto';
+}
+.matchColumn{
+    minWidth: '200px'
+}
+span{
+    float: 'left'; 
+    alignItems: 'center'
+}
