+++ src/UI/testPhoneInput/user-test-phone-input.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 import React, { useState } from 'react';
 import { Flex, Input, Typography } from 'antd';
 import { PhoneOutlined } from '@ant-design/icons';
