+++ src/UI/testPhoneInput/user-test-settings.jsx
@@ -12,6 +12,8 @@ import UserOrganizationInput from '@src/UI/user/user-organization-input.jsx'
 import Loader from '@components/loader/loader'
 import UserPasswordModal from './user-password-change'
 import ApiPath from '@components/enums.js'
+
+// Мой тестовый импорт
 import UserTestPhoneInput from '../testPhoneInput/user-test-phone-input'
 
 function UsersTestSettings() {
