+++ src/components/enums.js
@@ -1,4 +1,3 @@
-/* eslint-disable prettier/prettier */
 export const FILE_UPLOADING = {
   UPLOADING: 'uploading',
   DONE: 'done',
@@ -45,6 +44,10 @@ export const ROUTES = {
     TITLE: 'Групповой этап',
     PATH: (event_id, nomination_id) => `/judgment/${event_id}/${nomination_id}`,
   },
+  JUDGMENT_TIME_MATHCES: {
+    TITLE: 'Матчи на время',
+    PATH: (event_id, nomination_id) => `/test/${event_id}/${nomination_id}`,
+  },
 }
 
 export const ROUTER_ROUTES = {
@@ -62,7 +65,8 @@ export const ROUTER_ROUTES = {
   JUDGMENT: 'judgment/events',
   JUDGMENT_CREATE: 'create',
   JUDGMENT_EVENT_SETTINGS: ':eventID/settings',
-  JUDGMENT_GROUP_STAGE: ':judgment',
+  JUDGMENT_GROUP_STAGE: 'judgment/:event_id/:nomination_id',
+  JUDGMENT_TIME_MATCHES: 'test/:event_id/:nomination_id',
 }
 
 export const Locale = {
