+++ src/components/judgement/group-stage.jsx
@@ -2,7 +2,9 @@ import GroupTab from '../../UI/judgement/tab-group-stage'
 
 function GroupStage() {
   return (
+    <>
       <GroupTab />
+    </>
   )
 }
