+++ src/components/judgment/group-stage/group-stage.jsx
@@ -1,0 +1,7 @@
+import GroupTab from '../../../UI/judgment/group-stage/tab-group-stage'
+function GroupStage(){
+  return (
+      <GroupTab />
+  )
+}
+export default GroupStage
